This commit was generated by cvs2svn to compensate for changes in r16473, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16474 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/beos/make-bindist.sh
@@ -6,6 +6,12 @@
 # make-bindist.sh <strip> <vice-version> <zip|nozip> <top-srcdir>
 #                 $1      $2             $3         $4
 
+if [ ! -e src/x64 -o ! -e src/x128 -o ! -e src/xvic -o ! -e src/xpet -o ! -e src/xplus4 -o ! -e src/xcbm2 -o ! -e src/c1541 -o ! -e src/petcat -o ! -e src/cartconv ]
+then
+  echo Error: executable file\(s\) not found, do a \"make\" first
+  exit 1
+fi
+
 echo Generating BEOS port binary distribution.
 rm -f -r BeVICE-$2
 mkdir BeVICE-$2
